requirements.txt
Flask==3.0.0
Flask-CORS==3.0.10
gunicorn==20.1.0
google-auth==2.3.2
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.38.0
